% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinsku-package.R
\docType{package}
\name{addinsku-package}
\alias{addinsku}
\alias{addinsku-package}
\title{addinsku: Addinsku}
\description{
Kumpulan addins sederhana namun (harapannya dapat) bermanfaat.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aswansyahputra/addinsku}
  \item Report bugs at \url{https://github.com/aswansyahputra/addinsku/issues}
}

}
\author{
\strong{Maintainer}: Muhammad Aswan Syahputra \email{aswansyahputra@sensolution.id}

}
\keyword{internal}
